move thermal_pressure to plasmapy.formulary.misc
